chore(jans-cedarling): refactor, move authorize_result to authz module
